Progress - 32% - Success! OK!

+++ 4_Banco de Dados/Alura.Loja.Testes.ConsoleApp/Produto.cs
@@ -5,7 +5,7 @@
         public int Id { get; internal set; }
         public string Nome { get; internal set; }
         public string Categoria { get; internal set; }
-        public double Preco { get; internal set; }
+        public decimal Preco { get; internal set; }
 
         public override string ToString()
         {
